AndroidRunner integrated, VMAquirer deleted, Debugger functions added
