+++ src/processing/mode/android/AndroidRunner.java
@@ -22,14 +22,21 @@
 
 package processing.mode.android;
 
+import java.io.IOException;
 import java.io.PrintStream;
 import java.security.PublicKey;
 import java.util.Iterator;
 import java.util.List;
+import java.util.Map;
 import java.util.concurrent.*;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
+import com.sun.jdi.VirtualMachine;
+import com.sun.jdi.VirtualMachineManager;
+import com.sun.jdi.connect.AttachingConnector;
+import com.sun.jdi.connect.Connector;
+import com.sun.jdi.connect.IllegalConnectorArgumentsException;
 import jdk.nashorn.internal.runtime.linker.Bootstrap;
 import processing.app.ui.Editor;
 import processing.app.Messages;
@@ -46,6 +53,8 @@ public class AndroidRunner implements DeviceListener {
 
   protected PrintStream sketchErr;
   protected PrintStream sketchOut;
+
+  private VirtualMachine vm;
 
   public AndroidRunner(AndroidBuild build, RunnerListener listener) {
     this.build = build;
@@ -135,6 +144,43 @@ public class AndroidRunner implements DeviceListener {
     return status;
   }
 
+  public VirtualMachine connectVirtualMachine(int port) throws IOException {
+    String strPort = Integer.toString(port);
+    AttachingConnector connector = getConnector();
+    try {
+      vm = connect(connector, strPort);
+      return vm;
+    } catch (IllegalConnectorArgumentsException e) {
+      throw new IllegalStateException(e);
+    }
+  }
+
+  private AttachingConnector getConnector() {
+    VirtualMachineManager vmManager = org.eclipse.jdi.Bootstrap.virtualMachineManager();
+    for (Connector connector : vmManager.attachingConnectors()) {
+      if ("com.sun.jdi.SocketAttach".equals(connector.name())) {
+        return (AttachingConnector) connector;
+      }
+    }
+    throw new IllegalStateException();
+  }
+
+  private VirtualMachine connect(
+      AttachingConnector connector, String port) throws IllegalConnectorArgumentsException, IOException {
+    Map<String, Connector.Argument> args = connector
+        .defaultArguments();
+    Connector.Argument pidArgument = args.get("port");
+    if (pidArgument == null) {
+      throw new IllegalStateException();
+    }
+    pidArgument.setValue(port);
+
+    return connector.attach(args);
+  }
+
+  public VirtualMachine vm(){
+    return vm;
+  }
 
   private volatile Device lastRunDevice = null;
